+++ src/test/kotlin/devtests/Json.kt
@@ -1,10 +1,11 @@
 package devtests
 
+import java.io.Serializable
 import klib.extensions.toListOfType
 import klib.kLibInf
 import net.jemzart.jsonkraken.JsonArray
 
-data class User(var id: Int, val first_name: String, val last_name: String, val email: String)
+data class User(var id: Int, val first_name: String, val last_name: String, val email: String) : Serializable
 
 fun main() {
     val jsonData = "[{\n" +