+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - JsonKraken 2.0 dependency
 - `objectOutputStream` extension for File
 - `objectInputStream` extension for File
+- FFDB (FlatFile DataBase) (`klib.ffdb.FFDB`)
+- IncompatibleDatabaseException (`klib.exceptions.IncompatibleDatabaseException`)
 ### Changed
 - Upgraded Kotlin from 1.3.61 to 1.3.70
 ### Deprecated